+++ AGame/ResourceManager.cpp
@@ -338,10 +338,10 @@ void ResourceManager::DrawCursor()
 		int x, y;
 		AEGfxTextureSet(GetTexture("cursor"), 0.0f,0.0f);
 		AEInputGetCursorPosition(&x, &y);
-		x -= (float)AEGetWindowWidth() / 2.0f;
-		y -= (float)AEGetWindowHeight() / 2.0f;
+		x -= (int)((float)AEGetWindowWidth() / 2.0f);
+		y -= (int)((float)AEGetWindowHeight() / 2.0f);
 		y *= -1;
-		AEMtx33Trans(&trans, x, y);
+		AEMtx33Trans(&trans, (float)x, (float)y);
 		AEMtx33Scale(&scale, 100.0f, 100.0f);
 		AEMtx33Concat(&trans, &trans, &scale);
 		AEGfxSetTransform(trans.m);
@@ -354,21 +354,21 @@ void ResourceManager::CursorParticlesUpdate(const float& dt)
 {
 	int x, y;
 	AEInputGetCursorPosition(&x, &y);
-	x -= (float)AEGetWindowWidth() / 2.0f;
-	y -= (float)AEGetWindowHeight() / 2.0f;
+	x -= (int)((float)AEGetWindowWidth() / 2.0f);
+	y -= (int)((float)AEGetWindowHeight() / 2.0f);
 	y *= -1;
 	if (_cursor_particle_count > 0) {
 		AddCursorParticle();
-		_cursor_particles.back()._position.x = x;
-		_cursor_particles.back()._position.y = y;
+		_cursor_particles.back()._position.x = (float)x;
+		_cursor_particles.back()._position.y = (float)y;
 		--_cursor_particle_count;
 	}
 	AEGfxSetRenderMode(AEGfxRenderMode::AE_GFX_RM_TEXTURE);
 	AEGfxTextureSet(GetTexture("inkblob"), 0.0f, 0.0f);
 	for (auto& p : _cursor_particles) {
 		if (p._a <= 0.0f) {
-			p._position.x = x;
-			p._position.y = y;
+			p._position.x = (float)x;
+			p._position.y = (float)y;
 			p._dimension = (0.5f + AERandFloat()) * cursor_particle_scale;
 			p._scale = 1.0f;
 			p._a = 1.0f;
